more changes, void wvSetCharHandler(void (*proc)(wvParseStruct *,U16,U8)) is the big change, this lets you let a proc that will give you each char of the word document, with the letter in the U16 slot and the U8 set to CP1252 for 8bit mode, and not set to CP1252 for 16bit unicode. ill have a simple abiword importer that uses this very soon now.
git-svn-id: svn+ssh://svn.abisource.com/svnroot/wv/trunk@20248 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ table.c
@@ -537,11 +537,11 @@ void output_tablebg(pap *apap)
 
 	if (apap->ourtap.cell_backs[colcount] != 0)
 		{
-		fprintf(outputfile," bgcolor=\"#%.2x",(colorlookupr[apap->ourtap.cell_backs[colcount]])*shademodifier/100);
-		fprintf(outputfile,"%.2x",(colorlookupg[apap->ourtap.cell_backs[colcount]])*shademodifier/100);
-		fprintf(outputfile,"%.2x",(colorlookupb[apap->ourtap.cell_backs[colcount]])*shademodifier/100);
+		fprintf(outputfile," bgcolor=\"#%.2x",(int)((colorlookupr[apap->ourtap.cell_backs[colcount]])*shademodifier/100));
+		fprintf(outputfile,"%.2x",(int)((colorlookupg[apap->ourtap.cell_backs[colcount]])*shademodifier/100));
+		fprintf(outputfile,"%.2x",(int)((colorlookupb[apap->ourtap.cell_backs[colcount]])*shademodifier/100));
 		fprintf(outputfile,"\"");
-		sprintf(backgroundcolor,"#%.2x%.2x%.2x",(colorlookupr[apap->ourtap.cell_backs[colcount]])*shademodifier/100,(colorlookupg[apap->ourtap.cell_backs[colcount]])*shademodifier/100,(colorlookupb[apap->ourtap.cell_backs[colcount]])*shademodifier/100);
+		sprintf(backgroundcolor,"#%.2x%.2x%.2x",(int)((colorlookupr[apap->ourtap.cell_backs[colcount]])*shademodifier/100),(int)((colorlookupg[apap->ourtap.cell_backs[colcount]])*shademodifier/100),(int)((colorlookupb[apap->ourtap.cell_backs[colcount]])*shademodifier/100));
 		}
 
 	fprintf(outputfile,">\n");
